Fix broken shared library detection
It seems that lldb-3.8 is not broken on Ubuntu any more, at least for Ubuntu
16.04. Actually, this "workaround" prevents compilation on Ubuntu 16.04 and Arch
Linux.

It's tempting to remove the workaround straightforwardly, but it's unknown
whether the problem still exists on other distributions. To maintain
compatibility of previously broken packages while avoid the countereffect, this
commit searches through all LLVM libraries with heuristics.